requirements.txt
websocket-client>=1.6.0
fastapi>=0.100.0
uvicorn>=0.23.0
pycryptodome>=3.19.0
